refactor: move PDFMerger to pdf package and changed class modifier
